requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-openai
pypdf
chromadb
tiktoken
